+++ src/services/IEmployeeSystem.java
@@ -14,5 +14,5 @@ public interface IEmployeeSystem {
     void removeEmployeeByCode(int code, Employee manager) throws IllegalArgumentException, EmployeeSystemException; // remove employee by code
     Employee login(int code); // Authenticates an employee using a unique code
     boolean isManager(Employee employee); // Checks if the employee is a manager
-    void trackEmployeeHours(Employee employee, int hours); // Logs working hours
+    void trackEmployeeHours(Employee employee, double hours); // Logs working hours
 }
